backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.36

# NLP
nltk==3.9.1
spacy==3.7.6
sumy==0.11.0

# ML
scikit-learn==1.5.2
xgboost==2.1.3
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2

# Utilities
Werkzeug==3.1.3
python-dotenv==1.0.1
